requirements.txt
httpx>=0.28.0
flask>=3.0.0
websocket-client>=1.7.0
requests>=2.31.0
